+++ packages/auth-construct/src/index.ts
@@ -15,16 +15,5 @@ export {
   PhoneNumberLogin,
   TriggerEvent,
 } from './types.js';
-export {
-  AuthUserAttribute,
-  AuthStandardAttribute,
-  AuthCustomAttributeBase,
-  AuthCustomAttributeFactory,
-  AuthCustomBooleanAttribute,
-  AuthCustomDateTimeAttribute,
-  AuthCustomNumberAttribute,
-  AuthCustomStringAttribute,
-  Mutable,
-} from './attributes.js';
 export { AmplifyAuth } from './construct.js';
 export { triggerEvents } from './trigger_events.js';