+++ packages/cli/src/commands/pipeline-deploy/pipeline_deploy_command_factory.ts
@@ -5,6 +5,8 @@ import {
   PipelineDeployCommand,
   PipelineDeployCommandOptions,
 } from './pipeline_deploy_command.js';
+import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
+import { ClientConfigGeneratorAdapter } from '../../client-config/client_config_generator_adapter.js';
 
 /**
  * Creates pipeline deploy command
@@ -13,6 +15,10 @@ export const createPipelineDeployCommand = (): CommandModule<
   object,
   PipelineDeployCommandOptions
 > => {
+  const credentialProvider = fromNodeProviderChain();
+  const clientConfigGenerator = new ClientConfigGeneratorAdapter(
+    credentialProvider
+  );
   const backendDeployer = BackendDeployerFactory.getInstance();
-  return new PipelineDeployCommand(backendDeployer);
+  return new PipelineDeployCommand(clientConfigGenerator, backendDeployer);
 };